Fix default range for trigno
note: this changes get_channel_default_range to return a tuple of floats

+++ bomi/device_managers/protocols.py
@@ -67,7 +67,7 @@ class SupportsGetChannelMetadata(Protocol):
         """
 
     @staticmethod
-    def get_channel_default_range(channel: str) -> tuple[int, int]:
+    def get_channel_default_range(channel: str) -> tuple[float, float]:
         """
         Gets a reasonable range for the data of a given channel.
         """
