+++ bomi/device_managers/qtm_manager.py
@@ -40,7 +40,7 @@ class QtmDeviceManager(QObject):
                 raise ValueError("Not a valid QTM channel")
 
     @staticmethod
-    def get_channel_default_range(channel: str) -> tuple[int, int]:
+    def get_channel_default_range(channel: str) -> tuple[float, float]:
         match channel:
             case Channel.TORQUE:
                 return -60, 10
